requirements.txt
fastapi
uvicorn[standard]
python-multipart 
uvicorn 
onnxruntime 
pillow 
numpy
requests
